technology-xml-parser-r1-lm-dev/requirements.txt
pyspark
google-cloud-storage
pyyaml
pytest
